fix: moved sentTxStore test files location
