+++ src/features/pastelSentTxStore/__tests__/sentTxStore.test.ts
@@ -1,6 +1,6 @@
 import fs from 'fs'
 
-import { loadSentTxns } from '../../../features/pastelLib/sent-tx-store'
+import { loadSentTxns } from '../sent-tx-store'
 
 jest.mock('fs', () => ({
   promises: {
@@ -43,7 +43,6 @@ const mockTransactions = [
     vjoinsplit: [],
     size: 0,
     lastblock: '',
-    fee: undefined,
     type: '',
     detailedTxns: [
       {
